[BLD-41825] - refactored to ViewModel (instead of AndroidViewModel) + fixed a crush on lower APIs
