+++ SDKSamples/app/src/main/java/com/sdk/samples/SingletonSamplesViewModelFactory.kt
@@ -1,0 +1,48 @@
+package com.sdk.samples
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import java.lang.reflect.InvocationTargetException
+
+class SingletonSamplesViewModelFactory(vmInstance: SamplesViewModel) : ViewModelProvider.NewInstanceFactory() {
+
+    private val samplesViewModel: SamplesViewModel = vmInstance
+    private val viewModelFactory: MutableMap<Class<out ViewModel>, ViewModel> = mutableMapOf()
+
+    fun clear() {
+        viewModelFactory.clear()
+    }
+
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+
+        viewModelFactory[modelClass] = samplesViewModel
+
+        if (SamplesViewModel::class.java.isAssignableFrom(modelClass)) {
+
+            val shareVM: SamplesViewModel
+
+            if (viewModelFactory.containsKey(modelClass)) {
+                shareVM = viewModelFactory[modelClass] as SamplesViewModel
+            } else {
+                shareVM = try {
+                    modelClass.getConstructor(Runnable::class.java).newInstance(
+                        Runnable { viewModelFactory.remove(modelClass) }) as SamplesViewModel
+
+                } catch (e: Exception) {
+                    throw RuntimeException("Cannot create an instance of $modelClass", e)
+                } catch (e: IllegalAccessException) {
+                    throw RuntimeException("Cannot create an instance of $modelClass", e)
+                } catch (e: InstantiationException) {
+                    throw RuntimeException("Cannot create an instance of $modelClass", e)
+                } catch (e: InvocationTargetException) {
+                    throw RuntimeException("Cannot create an instance of $modelClass", e)
+                }
+                viewModelFactory[modelClass] = shareVM
+            }
+            return shareVM as T
+        }
+
+        return super.create(modelClass)
+    }
+
+}
