--- SDKSamples/app/src/main/res/anim/right_in_no_alpha.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-    <translate android:fromXDelta="100%" android:toXDelta="0" android:duration="500"/>
-</set>
