--- SDKSamples/app/src/main/res/layout/async_chat_config.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<androidx.core.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-    <!--<LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="14dp">
-
-        <androidx.appcompat.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:text="@string/apikey"
-                android:textColor="@color/colorPrimary"
-                android:textSize="16sp" />
-
-        <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/apiKey_edit"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:hint="@string/apikey"
-                android:lines="2"
-                android:textColorHint="@color/form_field_hint"
-                android:textSize="14sp" />
-
-        <androidx.appcompat.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:layout_marginTop="14dp"
-                android:text="@string/application_id_hint"
-                android:textColor="@color/colorPrimary"
-                android:textSize="16sp" />
-
-        <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/application_edit"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:hint="@string/application_id_hint"
-                android:lines="1"
-                android:textColorHint="@color/form_field_hint"
-                android:textSize="14sp" />
-
-        <androidx.appcompat.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:layout_marginTop="14dp"
-                android:text="@string/user_id_hint"
-                android:textColor="@color/colorPrimary"
-                android:textSize="16sp" />
-
-        <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/user_edit"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:hint="@string/user_id_hint"
-                android:inputType="numberDecimal"
-                android:lines="1"
-                android:textColorHint="@color/form_field_hint"
-                android:textSize="14sp" />
-
-        <androidx.appcompat.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:layout_marginTop="14dp"
-                android:text="@string/user_first"
-                android:textColor="@color/colorPrimary"
-                android:textSize="16sp" />
-
-        <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/user_first"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:hint="@string/user_first"
-                android:lines="1"
-                android:textColorHint="@color/form_field_hint"
-                android:textSize="14sp" />
-
-        <androidx.appcompat.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:layout_marginTop="14dp"
-                android:text="@string/user_last"
-                android:textColor="@color/colorPrimary"
-                android:textSize="16sp" />
-
-        <androidx.appcompat.widget.AppCompatEditText
-                android:id="@+id/user_last"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:hint="@string/user_last"
-                android:lines="1"
-                android:textColorHint="@color/form_field_hint"
-                android:textSize="14sp" />
-
-        <androidx.appcompat.widget.SwitchCompat
-                android:id="@+id/restore_switch"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="start"
-                android:layout_marginTop="14dp"
-                android:checked="true"
-                android:text="@string/restore_if_available" />
-
-        <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/start_button"
-                style="@style/main_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal"
-                android:layout_marginTop="20dp"
-                android:text="@string/start_async_chat" />
-
-    </LinearLayout>-->
-</androidx.core.widget.NestedScrollView>
