+++ SDKSamples/app/src/main/res/values-ldpi/dimens.xml
@@ -4,6 +4,5 @@
     <dimen name="chip_top_gap">5dp</dimen>
     <dimen name="chip_inner_padding">0dp</dimen>
     <dimen name="topic_titles_text">14sp</dimen>
-    <dimen name="main_button_text">10sp</dimen>
     <dimen name="buttons_text">8sp</dimen>
 </resources>
