+++ SDKSamples/app/src/main/res/values-mdpi/dimens.xml
@@ -4,7 +4,6 @@
     <dimen name="chip_top_gap">5dp</dimen>
     <dimen name="chip_inner_padding">0dp</dimen>
     <dimen name="topic_titles_text">14sp</dimen>
-    <dimen name="main_button_text">12sp</dimen>
     <dimen name="buttons_text">10sp</dimen>
 
 </resources>
