+++ SDKSamples/app/src/main/res/values-night/colors.xml
@@ -1,4 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="form_field_background">#55660000</color>
-</resources>
+<resources></resources>
