+++ SDKSamples/app/src/main/res/values/styles.xml
@@ -10,16 +10,6 @@
         <item name="android:windowBackground">@color/colorPrimaryDark</item>
     </style>
 
-    <style name="demo_chat_status_icon_style" partent="chat_textview_status_icon_style">
-        <item name="chat_status_ok">@drawable/send_confirm</item>
-        <item name="chat_status_pending">@drawable/send_error_2</item>
-        <item name="android:layout_marginLeft">2dp</item>
-        <item name="android:layout_marginBottom">2dp</item>
-        <item name="android:layout_height">10dp</item>
-        <item name="android:layout_width">18dp</item>
-        <item name="android:visibility">gone</item>
-    </style>
-
     <style name="CustomTextAppearance" parent="Base.TextAppearance.AppCompat.Large.Inverse">
         <item name="android:textSize">16sp</item>
         <item name="android:textColor">@color/form_field_hint</item>
